:art: Update `ServiceManager` to use `brew` when feature `homebrew` enabled.
