+++ src/service/macos/files/Library/LaunchDaemons/smartdns-rs.plist
@@ -2,26 +2,26 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>KeepAlive</key>
-	<true/>
-	<key>Label</key>
-	<string>smartdns-rs</string>
-	<key>LimitLoadToSessionType</key>
-	<array>
-		<string>Aqua</string>
-		<string>Background</string>
-		<string>LoginWindow</string>
-		<string>StandardIO</string>
-		<string>System</string>
-	</array>
-	<key>ProgramArguments</key>
-	<array>
-		<string>/usr/local/bin/smartdns</string>
-		<string>run</string>
-		<string>-c</string>
-		<string>/usr/local/etc/smartdns/smartdns.conf</string>
-	</array>
-	<key>KeepAlive</key>
-	<true/>
+        <key>KeepAlive</key>
+        <true/>
+        <key>Label</key>
+        <string>smartdns-rs</string>
+        <key>LimitLoadToSessionType</key>
+        <array>
+                <string>Aqua</string>
+                <string>Background</string>
+                <string>LoginWindow</string>
+                <string>StandardIO</string>
+                <string>System</string>
+        </array>
+        <key>ProgramArguments</key>
+        <array>
+                <string>/usr/local/sbin/smartdns</string>
+                <string>run</string>
+                <string>-c</string>
+                <string>/usr/local/etc/smartdns/smartdns.conf</string>
+        </array>
+        <key>RunAtLoad</key>
+        <true/>
 </dict>
 </plist>
